refactor: move API methods to entities folder
